Csw-t Refactoring Part 1 - Bundle all csw-t api files in package /features/cswt_api - Setup CswtApiConfig file with spring profile annotation - Remove redundant spring profile annotations - Combine CswtApi and CswtApiController - Leave controller in old api package
